restructure tooltip
