+++ packages/tooltip/src/Tooltip/index.ts
@@ -1,9 +1,10 @@
-import Tooltip, {
+import Tooltip from './Tooltip';
+import {
   Align,
   Justify,
   type TooltipProps,
   TriggerEvent,
-} from './Tooltip';
+} from './Tooltip.types';
 
 export { Align, Justify, type TooltipProps, TriggerEvent };
 export default Tooltip;
